Completely a hack job but shulker boxes render with shaders

+++ src/main/java/grondag/canvas/material/MaterialVertexFormat.java
@@ -42,15 +42,15 @@ public class MaterialVertexFormat {
 
 		int bytes = 0;
 		int count = 0;
-		
+
 		for (final MaterialVertextFormatElement e : elements) {
 			bytes += e.byteSize;
-		
+
 			if (e.attributeName != null) {
 				count++;
 			}
 		}
-		
+
 		attributeCount = count;
 		vertexStrideBytes = bytes;
 		vertexStrideInts = bytes / 4;
@@ -60,7 +60,7 @@ public class MaterialVertexFormat {
 	 * Enables generic vertex attributes and binds their location.
 	 * For use with non-VAO VBOs
 	 */
-	public void enableAndBindAttributes(int bufferOffset) {
+	public void enableAndBindAttributes(long bufferOffset) {
 		CanvasGlHelper.enableAttributes(attributeCount);
 		bindAttributeLocations(bufferOffset);
 	}
@@ -69,28 +69,28 @@ public class MaterialVertexFormat {
 	 * Enables generic vertex attributes and binds their location.
 	 * For use with non-VBO buffers.
 	 */
-	public void enableAndBindAttributes(ByteBuffer buffer, int bufferOffset) {
+	public void enableAndBindAttributes(ByteBuffer buffer, long bufferOffset) {
 		final int attributeCount = this.attributeCount;
 		CanvasGlHelper.enableAttributes(attributeCount);
 		int offset = 0;
 		int index = 1;
 		final int limit = elements.length;
-		
+
 		// NB: <= because element 0 is vertex
 		for(int i = 0; i <= attributeCount; i++) {
 			if(i < limit) {
 				final MaterialVertextFormatElement e = elements[i];
-				
+
 				if (e.attributeName != null) {
-					buffer.position(bufferOffset + offset);
-				
+					buffer.position((int) bufferOffset + offset);
+
 					if(Configurator.logGlStateChanges) {
 						CanvasMod.LOG.info(String.format("GlState: glVertexAttribPointer(%d, %d, %d, %b, %d, %s) [non-VBO]", index, e.elementCount, e.glConstant, e.isNormalized, vertexStrideBytes, buffer.toString()));
 					}
-					
+
 					GL20.glVertexAttribPointer(index++, e.elementCount, e.glConstant, e.isNormalized, vertexStrideBytes, buffer);
 				}
-				
+
 				offset += e.byteSize;
 			}
 		}
@@ -101,25 +101,25 @@ public class MaterialVertexFormat {
 	 * cases just call {@link #enableAndBindAttributes(int)}
 	 * @param attribCount How many attributes are currently enabled.  Any not in format should be bound to dummy index.
 	 */
-	public void bindAttributeLocations(int bufferOffset) {
+	public void bindAttributeLocations(long bufferOffset) {
 		int offset = 0;
 		int index = 1;
 		final int limit = elements.length;
 		final int attributeCount = this.attributeCount;
-		
+
 		// NB: <= because element 0 is vertex
 		for(int i = 0; i <= attributeCount; i++) {
 			if(i < limit) {
 				final MaterialVertextFormatElement e = elements[i];
-		
+
 				if (e.attributeName != null) {
 					if(Configurator.logGlStateChanges) {
 						CanvasMod.LOG.info(String.format("GlState: glVertexAttribPointer(%d, %d, %d, %b, %d, %d)", index, e.elementCount, e.glConstant, e.isNormalized, vertexStrideBytes, bufferOffset + offset));
 					}
-				
+
 					GL20.glVertexAttribPointer(index++, e.elementCount, e.glConstant, e.isNormalized, vertexStrideBytes, bufferOffset + offset);
 				}
-				
+
 				offset += e.byteSize;
 			}
 		}
